refactor. CQRS 패턴 해제
