+++ src/main/kotlin/io/runnershigh/backend/training/infrastructure/repository/TrainingPlanItemsRepository.kt
@@ -1,4 +1,4 @@
-package io.runnershigh.backend.training.infrastructure.repository.command
+package io.runnershigh.backend.training.infrastructure.repository
 
 import io.runnershigh.backend.training.infrastructure.entity.TrainingPlanItems
 import org.springframework.data.jpa.repository.JpaRepository
