+++ src/main/kotlin/io/runnershigh/backend/training/infrastructure/repository/TrainingSchedulesRepository.kt
@@ -1,0 +1,6 @@
+package io.runnershigh.backend.training.infrastructure.repository
+
+import io.runnershigh.backend.training.infrastructure.entity.TrainingSchedules
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface TrainingSchedulesRepository : JpaRepository<TrainingSchedules, Long>
