--- src/main/kotlin/io/runnershigh/backend/training/infrastructure/repository/command/TrainingSchedulesCommandRepository.kt
@@ -1,6 +1,0 @@
-package io.runnershigh.backend.training.infrastructure.repository.command
-
-import io.runnershigh.backend.training.infrastructure.entity.TrainingSchedules
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface TrainingSchedulesCommandRepository : JpaRepository<TrainingSchedules, Long>
