+++ src/main/kotlin/io/runnershigh/backend/user/infrastructure/repository/UserRepository.kt
@@ -1,8 +1,8 @@
-package io.runnershigh.backend.user.infrastructure.repository.query
+package io.runnershigh.backend.user.infrastructure.repository
 
 import io.runnershigh.backend.user.infrastructure.entity.UserEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface UserQueryRepository : JpaRepository<UserEntity, Int> {
+interface UserRepository : JpaRepository<UserEntity, Int> {
     fun findByLoginId(loginId: String): UserEntity?
 }
