--- src/test/kotlin/io/runnershigh/backend/BackendApplicationTests.kt
@@ -1,13 +1,0 @@
-package io.runnershigh.backend
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class BackendApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
