+++ src/test/kotlin/io/runnershigh/backend/training/application/command/TrainingSchedulesCommandUseCaseTest.kt
@@ -9,7 +9,7 @@ import io.runnershigh.backend.training.domain.mapper.toEntity
 import io.runnershigh.backend.training.exception.TrainingException
 import io.runnershigh.backend.training.exception.TrainingExceptionType
 import io.runnershigh.backend.training.infrastructure.entity.TrainingSchedules
-import io.runnershigh.backend.training.infrastructure.repository.command.TrainingSchedulesCommandRepository
+import io.runnershigh.backend.training.infrastructure.repository.TrainingSchedulesRepository
 import io.runnershigh.backend.training.ui.dto.SaveTrainingSchedule
 import io.runnershigh.backend.user.util.LoginUserContext
 import org.junit.jupiter.api.BeforeEach
@@ -24,7 +24,7 @@ import kotlin.test.assertEquals
 class TrainingSchedulesCommandUseCaseTest {
 
     @MockK
-    private lateinit var repository: TrainingSchedulesCommandRepository
+    private lateinit var repository: TrainingSchedulesRepository
 
     @MockK
     private lateinit var loginUserContext: LoginUserContext
